Ref: Created /card_top_details and moved file inside
