+++ app/src/components/common/index.ts
@@ -21,7 +21,7 @@ export { Select } from './form/select'
 export { Spinner } from './spinner'
 export { SubsectionTitle } from './text/subsection_title'
 
-export { SubsectionTitleWrapper } from './text/subsection_title_wrapper'
+export { SubsectionTitleWrapper } from '../market/sections/common_sections/card_top_details/subsection_title_wrapper'
 export { TD, TH, THead, TR, Table } from './tables/table'
 export { Textfield } from './form/textfield'
 export { SimpleTextfield } from './form/simiple_textfield'
