+++ app/src/components/market/sections/common_sections/card_top_details/market_top_details_closed/index.tsx
@@ -1,19 +1,22 @@
+import { BigNumber } from 'ethers/utils'
 import React, { useState } from 'react'
 import { useHistory } from 'react-router'
 import styled from 'styled-components'
 
-import { IMPORT_QUESTION_ID_KEY } from '../../../../common/constants'
-import { useConnectedWeb3Context } from '../../../../hooks'
-import { useGraphMarketsFromQuestion } from '../../../../hooks/useGraphMarketsFromQuestion'
-import { useWindowDimensions } from '../../../../hooks/useWindowDimensions'
-import theme from '../../../../theme'
-import { getContractAddress, getNativeAsset, getWrapToken } from '../../../../util/networks'
-import { getInitialCollateral, getMarketRelatedQuestionFilter, onChangeMarketCurrency } from '../../../../util/tools'
-import { MarketMakerData, MarketState, Token } from '../../../../util/types'
-import { SubsectionTitleWrapper } from '../../../common'
-import { MoreMenu } from '../../../common/form/more_menu'
+import { useConnectedWeb3Context } from '../../../../../../hooks'
+import { useGraphMarketsFromQuestion } from '../../../../../../hooks/useGraphMarketsFromQuestion'
+import { useWindowDimensions } from '../../../../../../hooks/useWindowDimensions'
+import theme from '../../../../../../theme'
+import { getContractAddress, getNativeAsset, getWrapToken } from '../../../../../../util/networks'
+import {
+  getInitialCollateral,
+  getMarketRelatedQuestionFilter,
+  onChangeMarketCurrency,
+} from '../../../../../../util/tools'
+import { MarketMakerData, MarketState, Token } from '../../../../../../util/types'
+import { SubsectionTitleWrapper } from '../../../../../common'
+import { CurrencySelector } from '../../../../common/currency_selector'
 import { AdditionalMarketData } from '../additional_market_data'
-import { CurrencySelector } from '../currency_selector'
 import { MarketData } from '../market_data'
 import { ProgressBar } from '../progress_bar'
 import { ProgressBarToggle } from '../progress_bar/toggle'
@@ -21,7 +24,6 @@ import { ProgressBarToggle } from '../progress_bar/toggle'
 const SubsectionTitleLeftWrapper = styled.div`
   display: flex;
   align-items: center;
-
   @media (max-width: ${props => props.theme.themeBreakPoints.sm}) {
     flex-grow: 1;
   }
@@ -35,22 +37,20 @@ const MarketCurrencySelector = styled(CurrencySelector)`
     min-width: auto;
   }
 `
-
 interface Props {
   marketMakerData: MarketMakerData
-  title?: string
+  collateral: BigNumber
 }
 
-const MarketTopDetailsOpen: React.FC<Props> = (props: Props) => {
+const MarketTopDetailsClosed: React.FC<Props> = (props: Props) => {
   const context = useConnectedWeb3Context()
   const { networkId, relay } = context
+  const { marketMakerData } = props
+  const history = useHistory()
+
   const { width } = useWindowDimensions()
   const isMobile = width <= parseInt(theme.themeBreakPoints.sm)
 
-  const [showingProgressBar, setShowingProgressBar] = useState(false)
-  const history = useHistory()
-
-  const { marketMakerData } = props
   const {
     address,
     answerFinalizedTimestamp,
@@ -65,52 +65,25 @@ const MarketTopDetailsOpen: React.FC<Props> = (props: Props) => {
     scaledLiquidityParameter,
     submissionIDs,
   } = marketMakerData
-
+  const { title } = question
   const ovmAddress = getContractAddress(networkId, 'omenVerifiedMarkets')
-  const creationDate = new Date(1000 * parseInt(creationTimestamp))
 
   const collateral = getInitialCollateral(networkId, marketMakerData.collateral, relay)
 
-  const currentTimestamp = new Date().getTime()
+  const [showingProgressBar, setShowingProgressBar] = useState(false)
+
+  const creationDate = new Date(1000 * parseInt(creationTimestamp))
 
   const formattedLiquidity: string = scaledLiquidityParameter ? scaledLiquidityParameter.toFixed(2) : '0'
 
   const isPendingArbitration = question.isPendingArbitration
   const arbitrationOccurred = question.arbitrationOccurred
 
-  const marketState =
-    question.resolution.getTime() > currentTimestamp
-      ? MarketState.open
-      : question.resolution.getTime() < currentTimestamp &&
-        (answerFinalizedTimestamp === null || answerFinalizedTimestamp.toNumber() * 1000 > currentTimestamp)
-      ? MarketState.finalizing
-      : isPendingArbitration
-      ? MarketState.arbitration
-      : answerFinalizedTimestamp && answerFinalizedTimestamp.toNumber() * 1000 < currentTimestamp
-      ? MarketState.closed
-      : MarketState.none
-
   const { markets: marketsRelatedQuestion } = useGraphMarketsFromQuestion(question.id)
 
   const toggleProgressBar = () => {
     setShowingProgressBar(!showingProgressBar)
   }
-
-  const moreMenuItems = [
-    {
-      onClick: () => {
-        localStorage.setItem(IMPORT_QUESTION_ID_KEY, question.id)
-        history.push('/create')
-      },
-      content: 'Add Currency',
-    },
-    {
-      onClick: () => {
-        toggleProgressBar()
-      },
-      content: showingProgressBar ? 'Hide Market State' : 'Show Market State',
-    },
-  ]
 
   const nativeAssetAddress = getNativeAsset(networkId, relay).address.toLowerCase()
   const wrapTokenAddress = getWrapToken(networkId).address.toLowerCase()
@@ -120,7 +93,7 @@ const MarketTopDetailsOpen: React.FC<Props> = (props: Props) => {
     <>
       <SubsectionTitleWrapper>
         <SubsectionTitleLeftWrapper>
-          {!relay && marketsRelatedQuestion.length > 1 && (
+          {marketsRelatedQuestion.length > 1 && (
             <MarketCurrencySelector
               addNativeAsset
               context={context}
@@ -136,14 +109,12 @@ const MarketTopDetailsOpen: React.FC<Props> = (props: Props) => {
           {(!isMobile || marketsRelatedQuestion.length === 1) && (
             <ProgressBarToggle
               active={showingProgressBar}
-              state={marketState}
+              state={'closed'}
               templateId={question.templateId}
               toggleProgressBar={toggleProgressBar}
             ></ProgressBarToggle>
           )}
         </SubsectionTitleLeftWrapper>
-
-        <MoreMenu items={marketsRelatedQuestion.length > 1 && isMobile ? moreMenuItems : new Array(moreMenuItems[0])} />
       </SubsectionTitleWrapper>
       {showingProgressBar && (
         <ProgressBar
@@ -153,14 +124,13 @@ const MarketTopDetailsOpen: React.FC<Props> = (props: Props) => {
           creationTimestamp={creationDate}
           pendingArbitration={isPendingArbitration}
           resolutionTimestamp={question.resolution}
-          state={marketState}
+          state={MarketState.closed}
         ></ProgressBar>
       )}
       <MarketData
         answerFinalizedTimestamp={marketMakerData.answerFinalizedTimestamp}
         collateralVolume={collateralVolume}
         currency={collateral}
-        isFinalize={marketState === MarketState.finalizing}
         lastActiveDay={lastActiveDay}
         liquidity={formattedLiquidity}
         resolutionTimestamp={question.resolution}
@@ -177,10 +147,10 @@ const MarketTopDetailsOpen: React.FC<Props> = (props: Props) => {
         oracle="Reality.eth"
         ovmAddress={ovmAddress}
         submissionIDs={submissionIDs}
-        title={question.title}
+        title={title}
       ></AdditionalMarketData>
     </>
   )
 }
 
-export { MarketTopDetailsOpen }
+export { MarketTopDetailsClosed }
