+++ app/src/components/market/sections/common_sections/card_top_details/more_menu/index.tsx
@@ -1,7 +1,7 @@
 import React, { DOMAttributes, useCallback, useState } from 'react'
 import styled, { css } from 'styled-components'
 
-import { IconMore } from '../../icons'
+import { IconMore } from '../../../../../common/icons'
 
 const MoreButtonWidth = '40px'
