+++ app/src/components/market/sections/common_sections/card_top_details/progress_bar/img/three_dots_1.tsx
