+++ app/src/components/market/sections/common_sections/card_top_details/progress_bar/toggle.tsx
@@ -1,9 +1,9 @@
 import React, { DOMAttributes } from 'react'
 import styled from 'styled-components'
 
-import { getMarketTitles } from '../../../../util/tools'
-import { MarketState } from '../../../../util/types'
-import { Button } from '../../../button'
+import { getMarketTitles } from '../../../../../../util/tools'
+import { MarketState } from '../../../../../../util/types'
+import { Button } from '../../../../../button'
 
 import { ThreeDots1 } from './img/three_dots_1'
 import { ThreeDots2 } from './img/three_dots_2'
