+++ app/src/components/market/sections/common_sections/card_top_details/subsection_title_wrapper/index.tsx
