+++ app/src/components/market/sections/market_profile/market_status/open.tsx
@@ -14,11 +14,11 @@ import { getUnit, isDust } from '../../../../../util/tools'
 import { BalanceItem, MarketDetailsTab, MarketMakerData, OutcomeTableValue } from '../../../../../util/types'
 import { Button, ButtonContainer } from '../../../../button'
 import { ButtonType } from '../../../../button/button_styling_types'
-import { MarketTopDetailsOpen } from '../../../common/market_top_details_open'
 import { OutcomeTable } from '../../../common/outcome_table'
 import { ViewCard } from '../../../common/view_card'
 import { WarningMessage } from '../../../common/warning_message'
 import { MarketScale } from '../../common_sections/card_bottom_details/market_scale'
+import { MarketTopDetailsOpen } from '../../common_sections/card_top_details/market_top_details_open'
 import { MarketBondContainer } from '../../market_bond/market_bond_container'
 import { MarketHistoryContainer } from '../../market_history/market_history_container'
 import { MarketNavigation } from '../../market_navigation'
